rename files correctly
