+++ services/QuillLMS/client/app/bundles/Teacher/components/content_hubs/coming_soon_tooltip.tsx
@@ -1,19 +1,14 @@
 import * as React from 'react'
 
-import ComingSoonTooltip from './coming_soon_tooltip'
+import { Tooltip, } from '../../../Shared/index'
 
-const QuillResourceLink = ({ text, href, }) => {
-  const className = "quill-button focus-on-light outlined medium grey"
-
-  if (href) {
-    return <a className={className} href={href} rel="noopener noreferrer" target="_blank">{text}</a>
-  }
-
+const ComingSoonTooltip = ({ tooltipTrigger, }) => {
   return (
-    <ComingSoonTooltip
-      tooltipTrigger={<button className={className} disabled={true}>{text}</button>}
+    <Tooltip
+      tooltipText="Coming soon!"
+      tooltipTriggerText={tooltipTrigger}
     />
   )
 }
 
-export default QuillResourceLink
+export default ComingSoonTooltip
